+++ examples/ephys_session.py
@@ -14,7 +14,7 @@ from aind_data_schema.core.session import (
     Stream,
     VisualStimulation,
 )
-from aind_data_schema.core.devices import Software
+from aind_data_schema.components.devices import Software
 from aind_data_schema.models.modalities import Modality
 
 session = Session(
