+++ examples/fip_ophys_rig.py
@@ -3,7 +3,7 @@
 from datetime import date, datetime, timezone
 
 import aind_data_schema.core.rig as r
-import aind_data_schema.core.devices as d
+import aind_data_schema.components.devices as d
 from aind_data_schema.models.modalities import Modality
 
 rig = r.Rig(
