+++ src/aind_data_schema/components/coordinates.py
