+++ src/aind_data_schema/components/stimulus.py
