+++ src/aind_data_schema/core/processing.py
@@ -6,7 +6,7 @@ from typing import List, Literal, Optional
 from pydantic import Field, ValidationInfo, field_validator
 
 from aind_data_schema.base import AindCoreModel, AindGeneric, AindGenericType, AindModel, AwareDatetimeWithDefault
-from aind_data_schema.core.tile import Tile
+from aind_data_schema.components.tile import Tile
 from aind_data_schema.models.process_names import ProcessName
